arca-collector/pyproject.toml
[project]
name = "arca-collector"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "playwright>=1.55.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
